Refactor next() and previous() slider methods
Instead of calculating the current value of 'left' property and the
current image width to set the new value of 'left' property, we make
use of index number as a multiplier to the image width to set the
new 'left' property.

This is especially useful to keep the slider working in case the user
zooms in and out on the page which would have led to discrepancies in
the current value of 'left' property and the current value of the image
width to be added when we used the initial method.

In short, initially we were using the existing value of 'left' property
to calculate the new value and now we are using only the actual width
of the image frame and calculating the 'left' property each time.

A fetchImageWidth() method has also been created to get the existing
width of the image frame.